YouTubeVideoTranscriber/requirements.txt
# python packages needed
python-dotenv
streamlit
youtube-transcript-api

# LLM APIs
google-generativeai

